Move utility functions into utils/
To organize this a bit better,
and not have all of them
in a huge file.
