+++ src/LocalCiProvider.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { Job } from './Job';
-import { getJobs, getRootPath } from './utils';
+import getJobs from './utils/getJobs';
+import getRootPath from './utils/getRootPath';
 
 export class LocalCiProvider implements vscode.TreeDataProvider<Job> {
   private _onDidChangeTreeData: vscode.EventEmitter<Job | undefined | void> =
@@ -16,7 +17,7 @@ export class LocalCiProvider implements vscode.TreeDataProvider<Job> {
 
   async getChildren(): Promise<Job[]> {
     return Promise.resolve(
-      (await getJobs(`${getRootPath()}/.circleci/config.yml`)).map(
+      getJobs(`${getRootPath()}/.circleci/config.yml`).map(
         (jobName) => new Job(jobName, vscode.TreeItemCollapsibleState.None)
       )
     );
