+++ src/utils/getCheckoutJobs.ts
@@ -1,0 +1,12 @@
+import getConfigFile from './getConfigFile';
+
+/** Gets the names of the jobs that have a 'checkout' step. */
+export default function getCheckoutJobs(inputFile: string): string[] {
+  const configFile = getConfigFile(inputFile);
+
+  return Object.keys(configFile?.jobs ?? []).filter((jobName) =>
+    configFile.jobs[jobName]?.steps?.some(
+      (step) => 'checkout' === step || step.checkout
+    )
+  );
+}
