+++ src/utils/getJobs.ts
@@ -1,0 +1,5 @@
+import getConfigFile from './getConfigFile';
+
+export default function getJobs(configFilePath: string): string[] | [] {
+  return Object.keys(getConfigFile(configFilePath)?.jobs ?? {});
+}
