+++ src/utils/getRootPath.ts
@@ -1,0 +1,8 @@
+import * as vscode from 'vscode';
+
+/** Gets the absolute path to this VS Code workspace. */
+export default function getRootPath(): string {
+  return vscode.workspace?.workspaceFolders?.length
+    ? vscode.workspace.workspaceFolders[0]?.uri?.path
+    : '';
+}
