+++ src/utils/getSpawnOptions.ts
@@ -1,0 +1,12 @@
+import getRootPath from './getRootPath';
+import getPath from './getPath';
+
+export default function getSpawnOptions(): Record<string, unknown> {
+  return {
+    cwd: getRootPath(),
+    env: {
+      ...process.env,
+      PATH: getPath(), // eslint-disable-line @typescript-eslint/naming-convention
+    },
+  };
+}
