+++ src/utils/index.d.ts
@@ -1,0 +1,27 @@
+interface Step {
+  checkout?: Record<string, unknown> | string;
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  attach_workspace?: {
+    at?: string;
+  };
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  persist_to_workspace?: {
+    root?: string;
+    paths?: Array<string>;
+  };
+  run?: {
+    command?: string;
+  };
+}
+
+interface ConfigFile {
+  jobs: Record<
+    string,
+    {
+      docker: Array<Record<string, string>>;
+      steps?: Array<Step>;
+      // eslint-disable-next-line @typescript-eslint/naming-convention
+      working_directory?: string;
+    }
+  >;
+}
